chore: change the structure of siamese network
