--- deeppavlov/models/ranking/siamese_embeddings_network.py
@@ -1,10 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-class SiameseEmbeddingsNetwork(metaclass=ABCMeta):
-
-    def __init__(self):
-        pass
-
-    @abstractmethod
-    def embeddings_model(self):
-        pass
